+++ src/pygaps/utilities/folder_utilities.py
@@ -1,7 +1,6 @@
 """General functions for stepping through folders."""
 
-import os.path
-
+from pathlib import Path
 from .exceptions import pgError
 
 
@@ -25,13 +24,6 @@ def util_get_file_paths(folder, extension=None):
     if extension is None:
         raise pgError("Must provide a file extension to look for")
 
-    paths = []
-
-    for root, _, files in os.walk(folder):
-        for file in files:
-            fullpath = os.path.join(root, file)
-            ext = os.path.splitext(fullpath)[-1].lower()
-            if ext == extension.lower():
-                paths.append(fullpath)
+    paths = Path(folder).rglob(f"*.{extension}")
 
     return paths