+++ tests/characterisation/test_initial_enthalpy.py
@@ -12,8 +12,6 @@ All pre-calculated data for characterisation can be found in the
 /.conftest file together with the other isotherm parameters.
 """
 
-import os
-
 import pytest
 from matplotlib.testing.decorators import cleanup
 from numpy import isclose
@@ -27,7 +25,6 @@ from .conftest import DATA_CALO_PATH
 @pytest.mark.characterisation
 class TestInitialEnthalpyPoint():
     """Test point initial enthalpy methods."""
-
     def test_ienthalpy_point_checks(self, basic_pointisotherm):
         """Checks for built-in safeguards."""
 
@@ -39,14 +36,14 @@ class TestInitialEnthalpyPoint():
     def test_ienthalpy_point(self, sample):
         """The point method."""
         sample = DATA_CALO[sample]
-        filepath = os.path.join(DATA_CALO_PATH, sample['file'])
+        filepath = DATA_CALO_PATH / sample['file']
         isotherm = pygaps.isotherm_from_jsonf(filepath)
 
-        ienth_poly = pygaps.initial_enthalpy_point(
-            isotherm, 'enthalpy').get('initial_enthalpy')
+        ienth_poly = pygaps.initial_enthalpy_point(isotherm, 'enthalpy'
+                                                   ).get('initial_enthalpy')
 
         err_relative = 0.1  # 10 percent
-        err_absolute = 1   #
+        err_absolute = 1  #
 
         assert isclose(ienth_poly, sample['ienth'], err_relative, err_absolute)
 
@@ -54,7 +51,7 @@ class TestInitialEnthalpyPoint():
     def test_ienthalpy_point_output(self):
         """Test verbosity."""
         sample = DATA_CALO['Takeda 5A']
-        filepath = os.path.join(DATA_CALO_PATH, sample['file'])
+        filepath = DATA_CALO_PATH / sample['file']
         isotherm = pygaps.isotherm_from_jsonf(filepath)
         pygaps.initial_enthalpy_point(isotherm, 'enthalpy', verbose=True)
 
@@ -62,7 +59,6 @@ class TestInitialEnthalpyPoint():
 @pytest.mark.characterisation
 class TestInitialEnthalpyFit():
     """Test fitting initial enthalpy methods."""
-
     def test_ienthalpy_comp_checks(self, basic_pointisotherm):
         """Checks for built-in safeguards."""
 
@@ -74,14 +70,14 @@ class TestInitialEnthalpyFit():
     def test_ienthalpy_comb(self, sample):
         """The combined polynomial method"""
         sample = DATA_CALO[sample]
-        filepath = os.path.join(DATA_CALO_PATH, sample['file'])
+        filepath = DATA_CALO_PATH / sample['file']
         isotherm = pygaps.isotherm_from_jsonf(filepath)
 
-        ienth_poly = pygaps.initial_enthalpy_comp(
-            isotherm, 'enthalpy').get('initial_enthalpy')
+        ienth_poly = pygaps.initial_enthalpy_comp(isotherm, 'enthalpy'
+                                                  ).get('initial_enthalpy')
 
         err_relative = 0.1  # 10 percent
-        err_absolute = 1   #
+        err_absolute = 1  #
 
         assert isclose(ienth_poly, sample['ienth'], err_relative, err_absolute)
 
@@ -89,6 +85,6 @@ class TestInitialEnthalpyFit():
     def test_ienthalpy_comb_output(self):
         """Test verbosity."""
         sample = DATA_CALO['Takeda 5A']
-        filepath = os.path.join(DATA_CALO_PATH, sample['file'])
+        filepath = DATA_CALO_PATH / sample['file']
         isotherm = pygaps.isotherm_from_jsonf(filepath)
         pygaps.initial_enthalpy_comp(isotherm, 'enthalpy', verbose=True)