+++ tests/characterisation/test_psd_meso.py
@@ -15,8 +15,6 @@ All pre-calculated data for characterisation can be found in
 the /.conftest file together with the other isotherm parameters.
 """
 
-import os
-
 import numpy as np
 import pytest
 from matplotlib.testing.decorators import cleanup
@@ -31,7 +29,6 @@ from .conftest import DATA_N77_PATH
 @pytest.mark.characterisation
 class TestPSDMeso():
     """Test mesopore size distribution calculation."""
-
     def test_psd_meso_checks(self, basic_pointisotherm):
         """Checks for built-in safeguards."""
 
@@ -63,29 +60,31 @@ class TestPSDMeso():
         # exclude datasets where it is not applicable
         if sample.get('psd_meso_pore_size', None):
 
-            filepath = os.path.join(DATA_N77_PATH, sample['file'])
+            filepath = DATA_N77_PATH / sample['file']
             isotherm = pygaps.isotherm_from_jsonf(filepath)
 
             result_dict = pmes.psd_mesoporous(
-                isotherm,
-                psd_model=method,
-                branch='des')
+                isotherm, psd_model=method, branch='des'
+            )
 
-            loc = np.where(result_dict['pore_distribution'] == max(result_dict['pore_distribution']))
+            loc = np.where(
+                result_dict['pore_distribution'] ==
+                max(result_dict['pore_distribution'])
+            )
             principal_peak = result_dict['pore_widths'][loc]
 
             err_relative = 0.05  # 5 percent
             err_absolute = 0.01  # 0.01
 
             assert np.isclose(
-                principal_peak,
-                sample['psd_meso_pore_size'],
-                err_relative, err_absolute)
+                principal_peak, sample['psd_meso_pore_size'], err_relative,
+                err_absolute
+            )
 
     @cleanup
     def test_psd_meso_verbose(self):
         """Test verbosity."""
-        data = DATA['MCM-41']
-        filepath = os.path.join(DATA_N77_PATH, data['file'])
+        sample = DATA['MCM-41']
+        filepath = DATA_N77_PATH / sample['file']
         isotherm = pygaps.isotherm_from_jsonf(filepath)
         pygaps.psd_mesoporous(isotherm, verbose=True)