+++ tests/core/test_modelisotherm.py
@@ -16,7 +16,6 @@ from ..characterisation.conftest import DATA_N77_PATH
 @pytest.mark.core
 class TestModelIsotherm():
     """Test the ModelIsotherm class."""
-
     def test_isotherm_create(self):
         """Check isotherm can be created from basic data."""
         isotherm_param = {
@@ -46,9 +45,7 @@ class TestModelIsotherm():
 
         # regular creation, DataFrame
         pygaps.ModelIsotherm(
-            isotherm_data=isotherm_data,
-            model='Henry',
-            **isotherm_param
+            isotherm_data=isotherm_data, model='Henry', **isotherm_param
         )
 
         # regular creation, desorption
@@ -70,24 +67,17 @@ class TestModelIsotherm():
         # Missing pressure/loading
         with pytest.raises(pygaps.ParameterError):
             pygaps.ModelIsotherm(
-                pressure=pressure,
-                loading=None,
-                **isotherm_param
+                pressure=pressure, loading=None, **isotherm_param
             )
 
         # Missing model
         with pytest.raises(pygaps.ParameterError):
-            pygaps.ModelIsotherm(
-                isotherm_data=isotherm_data,
-                **isotherm_param
-            )
+            pygaps.ModelIsotherm(isotherm_data=isotherm_data, **isotherm_param)
 
         # Wrong model
         with pytest.raises(pygaps.ParameterError):
             pygaps.ModelIsotherm(
-                isotherm_data=isotherm_data,
-                model='Wrong',
-                **isotherm_param
+                isotherm_data=isotherm_data, model='Wrong', **isotherm_param
             )
 
         # Wrong branch
@@ -112,8 +102,10 @@ class TestModelIsotherm():
         model = pygaps.modelling.get_isotherm_model('Henry')
         pygaps.ModelIsotherm(
             model=model,
-            material='Test', material_batch='Henry',
-            temperature=303, adsorbate='nitrogen'
+            material='Test',
+            material_batch='Henry',
+            temperature=303,
+            adsorbate='nitrogen'
         )
 
     def test_isotherm_create_from_isotherm(self, basic_isotherm):
@@ -137,23 +129,27 @@ class TestModelIsotherm():
         )
 
     @cleanup
-    @pytest.mark.parametrize('file, ',
-                             [(data['file']) for data in list(DATA.values())])
+    @pytest.mark.parametrize(
+        'file, ', [(data['file']) for data in list(DATA.values())]
+    )
     def test_isotherm_create_guess(self, file):
         """Check isotherm can be guessed from PointIsotherm."""
 
-        filepath = os.path.join(DATA_N77_PATH, file)
+        filepath = DATA_N77_PATH / file
         isotherm = pygaps.isotherm_from_jsonf(filepath)
 
         pygaps.ModelIsotherm.from_pointisotherm(
-            isotherm, guess_model='all', verbose=True)
+            isotherm, guess_model='all', verbose=True
+        )
 
         pygaps.ModelIsotherm.from_pointisotherm(
-            isotherm, guess_model=['Henry', 'Langmuir'], verbose=True)
+            isotherm, guess_model=['Henry', 'Langmuir'], verbose=True
+        )
 
         with pytest.raises(pygaps.ParameterError):
             pygaps.ModelIsotherm.from_pointisotherm(
-                isotherm, guess_model=['Henry', 'DummyModel'], verbose=True)
+                isotherm, guess_model=['Henry', 'DummyModel'], verbose=True
+            )
 
     def test_isotherm_ret_pressure(self, basic_modelisotherm, use_adsorbate):
         """Check that all the functions in ModelIsotherm return their specified parameter."""
@@ -163,24 +159,34 @@ class TestModelIsotherm():
             basic_modelisotherm.pressure(branch='des')
 
         # Regular return
-        assert numpy.array_equal(basic_modelisotherm.pressure(5), [1.0, 2.25, 3.5, 4.75, 6.0])
+        assert numpy.array_equal(
+            basic_modelisotherm.pressure(5), [1.0, 2.25, 3.5, 4.75, 6.0]
+        )
 
         # Unit specified
-        assert numpy.array_equal(basic_modelisotherm.pressure(5, pressure_unit='Pa'), [100000, 225000, 350000, 475000, 600000])
+        assert numpy.array_equal(
+            basic_modelisotherm.pressure(5, pressure_unit='Pa'),
+            [100000, 225000, 350000, 475000, 600000]
+        )
 
         # Mode specified
-        assert basic_modelisotherm.pressure(5, pressure_mode='relative')[
-            0] == pytest.approx(0.12849, 0.001)
+        assert basic_modelisotherm.pressure(
+            5, pressure_mode='relative'
+        )[0] == pytest.approx(0.12849, 0.001)
 
         # Range specified
-        assert numpy.array_equal(basic_modelisotherm.pressure(5, limits=(2, 5)), [2.25, 3.5, 4.75])
+        assert numpy.array_equal(
+            basic_modelisotherm.pressure(5, limits=(2, 5)), [2.25, 3.5, 4.75]
+        )
 
         # Indexed option specified
-        assert basic_modelisotherm.pressure(5, indexed=True).equals(pandas.Series(
-            [1.0, 2.25, 3.5, 4.75, 6.0]
-        ))
+        assert basic_modelisotherm.pressure(5, indexed=True).equals(
+            pandas.Series([1.0, 2.25, 3.5, 4.75, 6.0])
+        )
 
-    def test_isotherm_ret_loading(self, basic_modelisotherm, use_material, use_adsorbate):
+    def test_isotherm_ret_loading(
+        self, basic_modelisotherm, use_material, use_adsorbate
+    ):
         """Check that all the functions in ModelIsotherm return their specified parameter."""
 
         # Wrong branch
@@ -191,29 +197,37 @@ class TestModelIsotherm():
         assert basic_modelisotherm.loading(5)[0] == pytest.approx(1.0, 1e-5)
 
         # Loading unit specified
-        assert basic_modelisotherm.loading(
-            5, loading_unit='mol')[0] == pytest.approx(0.001, 1e-5)
+        assert basic_modelisotherm.loading(5, loading_unit='mol'
+                                           )[0] == pytest.approx(0.001, 1e-5)
 
         # Loading basis specified
-        assert numpy.isclose(basic_modelisotherm.loading(
-            5, loading_basis='mass', loading_unit='g')[0], 0.02801, 1e-4, 1e-4)
+        assert numpy.isclose(
+            basic_modelisotherm.loading(
+                5, loading_basis='mass', loading_unit='g'
+            )[0], 0.02801, 1e-4, 1e-4
+        )
 
         # Adsorbent unit specified
-        assert basic_modelisotherm.loading(5, adsorbent_unit='kg')[
-            0] == pytest.approx(1000, 1e-5)
+        assert basic_modelisotherm.loading(5, adsorbent_unit='kg'
+                                           )[0] == pytest.approx(1000, 1e-5)
 
         # Adsorbent basis specified
-        assert basic_modelisotherm.loading(5, adsorbent_basis='volume',
-                                           adsorbent_unit='cm3')[0] == pytest.approx(10, 1e-5)
+        assert basic_modelisotherm.loading(
+            5, adsorbent_basis='volume', adsorbent_unit='cm3'
+        )[0] == pytest.approx(10, 1e-5)
 
         # Range specified
-        assert basic_modelisotherm.loading(5, limits=(2, 5))[0] == pytest.approx(2.25, 1e-5)
+        assert basic_modelisotherm.loading(5, limits=(2, 5)
+                                           )[0] == pytest.approx(2.25, 1e-5)
 
         # Indexed option specified
-        assert isinstance(basic_modelisotherm.loading(
-            5, indexed=True), pandas.Series)
+        assert isinstance(
+            basic_modelisotherm.loading(5, indexed=True), pandas.Series
+        )
 
-    def test_isotherm_ret_loading_at(self, basic_modelisotherm, use_material, use_adsorbate):
+    def test_isotherm_ret_loading_at(
+        self, basic_modelisotherm, use_material, use_adsorbate
+    ):
         """Check that all the functions in ModelIsotherm return their specified parameter."""
 
         # Wrong branch
@@ -226,12 +240,14 @@ class TestModelIsotherm():
 
         # Pressure unit specified
         loading_punit = basic_modelisotherm.loading_at(
-            100000, pressure_unit='Pa')
+            100000, pressure_unit='Pa'
+        )
         assert loading_punit == pytest.approx(1.0, 1e-5)
 
         # Pressure mode specified
         loading_mode = basic_modelisotherm.loading_at(
-            0.5, pressure_mode='relative')
+            0.5, pressure_mode='relative'
+        )
         assert loading_mode == pytest.approx(3.89137, 1e-4)
 
         # Loading unit specified
@@ -240,21 +256,24 @@ class TestModelIsotherm():
 
         # Loading basis specified
         loading_lbasis = basic_modelisotherm.loading_at(
-            1, loading_basis='mass', loading_unit='g')
+            1, loading_basis='mass', loading_unit='g'
+        )
         # Loading basis specified
         assert numpy.isclose(loading_lbasis, 0.02801, 1e-4, 1e-4)
 
         # Adsorbent unit specified
-        loading_bunit = basic_modelisotherm.loading_at(
-            1, adsorbent_unit='kg')
+        loading_bunit = basic_modelisotherm.loading_at(1, adsorbent_unit='kg')
         assert loading_bunit == pytest.approx(1000, 0.1)
 
         # Adsorbent basis specified
         loading_bads = basic_modelisotherm.loading_at(
-            1, adsorbent_basis='volume', adsorbent_unit='cm3')
+            1, adsorbent_basis='volume', adsorbent_unit='cm3'
+        )
         assert loading_bads == pytest.approx(10, 1e-3)
 
-    def test_isotherm_ret_pressure_at(self, basic_modelisotherm, use_material, use_adsorbate):
+    def test_isotherm_ret_pressure_at(
+        self, basic_modelisotherm, use_material, use_adsorbate
+    ):
         """Check that all the functions in ModelIsotherm return their specified parameter."""
 
         # Wrong branch
@@ -267,35 +286,43 @@ class TestModelIsotherm():
 
         # Pressure unit specified
         pressure_punit = basic_modelisotherm.pressure_at(
-            1.0, pressure_unit='Pa')
+            1.0, pressure_unit='Pa'
+        )
         assert pressure_punit == pytest.approx(100000, 1)
 
         # Pressure mode specified
         pressure_mode = basic_modelisotherm.pressure_at(
-            3.89137, pressure_mode='relative')
+            3.89137, pressure_mode='relative'
+        )
         assert pressure_mode == pytest.approx(0.5, 1e-5)
 
         # Loading unit specified
         pressure_lunit = basic_modelisotherm.pressure_at(
-            0.001, loading_unit='mol')
+            0.001, loading_unit='mol'
+        )
         assert pressure_lunit == pytest.approx(1, 1e-5)
 
         # Loading basis specified
         pressure_lbasis = basic_modelisotherm.pressure_at(
-            0.02808, loading_basis='mass', loading_unit='g')
+            0.02808, loading_basis='mass', loading_unit='g'
+        )
         assert pressure_lbasis == pytest.approx(1, 1e-2)
 
         # Adsorbent unit specified
         pressure_bunit = basic_modelisotherm.pressure_at(
-            1000, adsorbent_unit='kg')
+            1000, adsorbent_unit='kg'
+        )
         assert pressure_bunit == pytest.approx(1.0, 1e-5)
 
         # Adsorbent basis specified
         pressure_bads = basic_modelisotherm.pressure_at(
-            10, adsorbent_basis='volume', adsorbent_unit='cm3')
+            10, adsorbent_basis='volume', adsorbent_unit='cm3'
+        )
         assert pressure_bads == pytest.approx(1.0, 1e-5)
 
-    def test_isotherm_spreading_pressure_at(self, basic_modelisotherm, use_adsorbate):
+    def test_isotherm_spreading_pressure_at(
+        self, basic_modelisotherm, use_adsorbate
+    ):
         """Check that all the functions in ModelIsotherm return their specified parameter."""
 
         # Wrong branch
@@ -308,12 +335,14 @@ class TestModelIsotherm():
 
         # Pressure unit specified
         spressure_punit = basic_modelisotherm.spreading_pressure_at(
-            100000, pressure_unit='Pa')
+            100000, pressure_unit='Pa'
+        )
         assert spressure_punit == pytest.approx(1.0, 1e-5)
 
         # Mode specified
         spressure_mode = basic_modelisotherm.spreading_pressure_at(
-            0.5, pressure_mode='relative')
+            0.5, pressure_mode='relative'
+        )
         assert spressure_mode == pytest.approx(3.89137, 1e-2)
 
     @cleanup