+++ tests/parsing/test_bel.py
@@ -9,7 +9,6 @@ from .conftest import DATA_BEL
 
 @pytest.mark.parsing
 class TestBEL():
-
     def test_read_bel(self):
         """Tests reading of a bel data file"""
 
@@ -17,7 +16,7 @@ class TestBEL():
 
             isotherm = pygaps.isotherm_from_bel(path=path)
 
-            json_path = path.replace('.DAT', '.json')
+            json_path = str(path).replace('.DAT', '.json')
 
             with open(json_path, 'r') as file:
                 assert isotherm == pygaps.isotherm_from_json(file.read())