+++ tests/utilities/test_utilities.py
@@ -2,30 +2,30 @@
 Tests utilities
 """
 
-import os
+from pathlib import Path
 
 import pytest
 
 import pygaps.utilities as utilities
 
-HERE = os.path.dirname(__file__)
-
 
 @pytest.mark.core
 def test_matplotlib_chemformula():
     assert utilities.string_utilities.convert_chemformula(
-        "C4H10") == "$C_{4}H_{10}$"
+        "C4H10"
+    ) == "$C_{4}H_{10}$"
 
 
 @pytest.mark.core
 def test_file_paths():
-    path = os.path.join(HERE, 'tst')
+    path = Path(__file__) / 'tst'
 
     with pytest.raises(Exception):
         utilities.folder_utilities.util_get_file_paths(path)
 
-    known_paths = [os.path.join(path, '1.tst'), os.path.join(path, '2.tst')]
+    known_paths = [path / '1.tst', path / '2.tst']
     paths = utilities.folder_utilities.util_get_file_paths(
-        path, extension='.tst')
+        path, extension='tst'
+    )
 
     assert all([path in known_paths for path in paths])